plugin: wire in plugin loader for node creation
License: MIT
